Fix the codegen issue
This one requires a bit more explanation. It's a bug that's been driving
me insane for the whereabouts of two months.

I've been seeing a codegen difference between the "original" exe and the
metroskrew in the following code:

    int func(int param)
    {
        switch (param) {
        case 10:
        case 15:
        case 16:
        case 19:
            return 1;
        }
        return 0;
    }

One version would generate a jump table, the other version would
generate a tree of if/else statements. The interesting thing, is that
when 19 was changed to 20, both versions would generate an if/else tree,
and if it was changed to 18, both versions would generate a jumptable.

At first, I blamed the assembly code that metroskrew was implemented it,
so I rewrote it in C. This did not fix it.

Then I blamed my implementation of the Win32 syscalls, so I ported
metroskrew to run natively on windows (I did not find another way to
selectively pick the WINE implementation or my own), and debugged my
syscalls. This did not fix it.

Then I started stripping the project down to the bare basics, I'd just
take the executable, relink it against the native windows functions, and
set the entrypoint. This is where I noticed something.

Even if I used my own implementations of the syscalls, the behavior
would only match the original if the original entrypoint was used,
rather than a new startfile and main() function.

Setting up my project to link with a custom startfile modified from the
mingw-w64 project, I noticed that commenting out the _fpreset()
function[1] caused the behavior to match. This function is implemented
in msvcrt.dll[2]. After messing with this function for a bit I found my
answer.

It turns out, by default, windows will initialize the FPU control word
with the value 0x027f, but Linux, and any modern C runtime for windows
(or at least mingw-w64), will end up setting it to 0x037f. This causes
the floating point instructions to behave slightly differently, and
breaks legacy programs.

I'm really happy to have figured this out, finally. I can already feel
some microsoft employee laughing at my struggles, but I finally did it,
damn it.

[1]: https://github.com/mirror/mingw-w64/blob/b2301d20800fb9afc853d7188dd2e0fa8621d1f9/mingw-w64-crt/crt/crtexe.c#L263
[2]: https://github.com/wine-mirror/wine/blob/2fc773199080dedbe64a33ef66be02a4f44bcd08/dlls/msvcrt/math.c#L1546-L1563